Support `/metrics/find` API, support multi-backend requests
Unfortunately this is something of a big-bang update. In the future
I will try to stick to iterative commits. This commit adds a few important
features:

If you had tried to use metaphite with something like grafana, you
would see that it does not autocomplete your choices when constructing
queries. That was because this API was not supported. With this update,
metaphite proxies its requests to all matching backends, merges the results,
and sends it to the client.

Popular dashboard systems such as grafana bundle all queries for a panel into
a single POST request. This means that a single POST request could have target
expressions that reference different backends. To handle this, metaphite groups
the target expressions by backend, proxies the request to matching backends,
and merges the results before sending them to the client.

This is currently quite rough. For instance, metaphite will wait for all backends
to respond (or time out) before sending any data to the client. And it will load
the entire result into memory instead of streaming it to the client. This can be
improved if it becomes a problem.